+++ src/components/SignIn/SignIn.js
@@ -2,8 +2,8 @@ import React, { Component } from "react";
 import { withRouter } from "react-router-dom";
 
 import { SignUpLink } from "../SignUp";
-import { withFirebase } from "../../Firebase";
-import * as ROUTES from "../../../constants/routes";
+import { withFirebase } from "../Firebase";
+import * as ROUTES from "../../constants/routes";
 
 const SignInPage = () => (
   <div>
