+++ src/components/SignOut/SignOut.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { withFirebase } from "../../Firebase";
+import { withFirebase } from "../Firebase";
 
 const SignOutButton = ({ firebase }) => (
   <button type="button" onClick={firebase.doSignOut}>
