+++ src/components/SignUp/SignUp.js
@@ -1,8 +1,8 @@
 import React, { Component } from "react";
 import { Link, withRouter } from "react-router-dom";
 
-import { withFirebase } from "../../Firebase";
-import * as ROUTES from "../../../constants/routes";
+import { withFirebase } from "../Firebase";
+import * as ROUTES from "../../constants/routes";
 
 const SignUpPage = () => (
   <div>
